fix: use last AssessmentStep for LogicalLocation
Eddie correctly pointed out that the location is for the entire evaluation,
not just where it started. Using the first step was misleading.

Changed from log.Steps[0] to log.Steps[len(log.Steps)-1] to use the
last step that was run, which better represents the final evaluation state.

Addresses review feedback from PR #171

Signed-off-by: Zohayb Bhatti <zobi@revanite.io>